Fire queue for second meet.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Core/LauncherHardware.java
@@ -67,7 +67,7 @@ public class LauncherHardware {
         return distance * distanceMultiplier;
     }
 
-    public void readyFire(int speedTarget) {
+    public void readyFire(double speedTarget) {
         waitingToFire = true;
         setLauncherSpeed(speedTarget);
     }
